@@ -57,12 +57,20 @@ public static class NullCoalescingOperator
         List<int> numbers = null;
         int? a = null;
 
+        Console.WriteLine((numbers is null)); // expected: true
+        // if numbers is null, initialize it. Then, add 5 to numbers
         (numbers ??= new List<int>()).Add(5);
         Console.WriteLine(string.Join(" ", numbers));  // output: 5
+        Console.WriteLine((numbers is null)); // expected: false        
 
+        
+        Console.WriteLine((a is null)); // expected: true
+        Console.WriteLine((a ?? 3)); // expected: 3 since a is still null 
+        // if a is null then assign 0 to a and add a to the list
         numbers.Add(a ??= 0);
+        Console.WriteLine((a is null)); // expected: false        
         Console.WriteLine(string.Join(" ", numbers));  // output: 5 0
-        Console.WriteLine(a);  // output: 0
+        Console.WriteLine(a);  // output: 0	        
         // </SnippetAssignment>
     }
 }
