@@ -8,29 +8,29 @@ namespace DynamicExamples
         static void Main(string[] args)
         {
             ExampleClass ec = new ExampleClass();
-            Console.WriteLine(ec.exampleMethod(10));
-            Console.WriteLine(ec.exampleMethod("value"));
+            Console.WriteLine(ec.ExampleMethod(10));
+            Console.WriteLine(ec.ExampleMethod("value"));
 
-            // The following line causes a compiler error because exampleMethod
+            // The following line causes a compiler error because ExampleMethod
             // takes only one argument.
-            //Console.WriteLine(ec.exampleMethod(10, 4));
+            //Console.WriteLine(ec.ExampleMethod(10, 4));
 
             dynamic dynamic_ec = new ExampleClass();
-            Console.WriteLine(dynamic_ec.exampleMethod(10));
+            Console.WriteLine(dynamic_ec.ExampleMethod(10));
 
-            // Because dynamic_ec is dynamic, the following call to exampleMethod
+            // Because dynamic_ec is dynamic, the following call to ExampleMethod
             // with two arguments does not produce an error at compile time.
             // However, it does cause a run-time error.
-            //Console.WriteLine(dynamic_ec.exampleMethod(10, 4));
+            //Console.WriteLine(dynamic_ec.ExampleMethod(10, 4));
         }
     }
 
     class ExampleClass
     {
-        static dynamic field;
-        dynamic prop { get; set; }
+        static dynamic _field;
+        dynamic Prop { get; set; }
 
-        public dynamic exampleMethod(dynamic d)
+        public dynamic ExampleMethod(dynamic d)
         {
             dynamic local = "Local variable";
             int two = 2;
